+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 0.5alpha
 ----------------
+- renames: dl_link_control_options -> link_control_options, dl_chattr() -> modlco(), dl_column_options -> dl_attlco
 - bugfixes
 
 Version 0.4